+++ x509-certificate/CHANGELOG.rst
@@ -8,6 +8,7 @@
 (Not yet released)
 
 * Canonical home of project moved to https://github.com/indygreg/cryptography-rs.
+* chrono crate upgraded from 0.4.22 to 0.4.23.
 
 0.15.0
 ======